Feat #5 : Engagement 엔티티 설정
Engagement (생성자/Getter/Table/Entity)
- id
- schedule
- attendee
- createdAt
- requestStatus
